backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-multipart
opencv-python-headless
numpy
tenacity
python-dotenv
psycopg2-binary
sqlalchemy
